Fix Parameterized tests on MsTestV2 integration.
