+++ src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Testing/MsTestV2/TestMethodAttributeExecuteIntegration.cs
@@ -1,7 +1,9 @@
 using System;
+using Datadog.Trace.Ci;
 using Datadog.Trace.ClrProfiler.CallTarget;
 using Datadog.Trace.Configuration;
 using Datadog.Trace.DuckTyping;
+using Datadog.Trace.Logging;
 
 namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing.MsTestV2
 {
@@ -21,6 +23,7 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing.MsTestV2
     {
         private const string IntegrationName = nameof(IntegrationIds.MsTestV2);
         private static readonly IntegrationInfo IntegrationId = IntegrationRegistry.GetIntegrationInfo(IntegrationName);
+        private static readonly IDatadogLogger Log = DatadogLogging.GetLoggerFor(typeof(TestMethodAttributeExecuteIntegration));
 
         /// <summary>
         /// OnMethodBegin callback
@@ -31,8 +34,15 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing.MsTestV2
         /// <param name="testMethod">Test method instance</param>
         /// <returns>Calltarget state value</returns>
         public static CallTargetState OnMethodBegin<TTarget, TTestMethod>(TTarget instance, TTestMethod testMethod)
+            where TTestMethod : ITestMethod, IDuckType
         {
-            return CallTargetState.GetDefault();
+            if (!Common.TestTracer.Settings.IsIntegrationEnabled(IntegrationId))
+            {
+                return CallTargetState.GetDefault();
+            }
+
+            var scope = MsTestIntegration.OnMethodBegin(testMethod, testMethod.Type);
+            return new CallTargetState(scope);
         }
 
         /// <summary>
@@ -57,14 +67,42 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing.MsTestV2
                     {
                         object testResultObject = returnValueArray.GetValue(0);
                         if (testResultObject != null &&
-                            testResultObject.TryDuckCast<TestResultStruct>(out var testResult) &&
-                            testResult.TestFailureException != null)
+                            testResultObject.TryDuckCast<TestResultStruct>(out var testResult))
                         {
-                            Exception testException = testResult.TestFailureException.InnerException ?? testResult.TestFailureException;
-                            string testExceptionName = testException.GetType().Name;
-                            if (testExceptionName != "UnitTestAssertException" && testExceptionName != "AssertInconclusiveException")
+                            string errorMessage = null;
+                            string errorStackTrace = null;
+
+                            if (testResult.TestFailureException != null)
                             {
-                                scope.Span.SetException(testException);
+                                Exception testException = testResult.TestFailureException.InnerException ?? testResult.TestFailureException;
+                                string testExceptionName = testException.GetType().Name;
+                                if (testExceptionName != "UnitTestAssertException" && testExceptionName != "AssertInconclusiveException")
+                                {
+                                    scope.Span.SetException(testException);
+                                }
+
+                                errorMessage = testException.Message;
+                                errorStackTrace = testException.StackTrace;
+                            }
+
+                            switch (testResult.Outcome)
+                            {
+                                case UnitTestOutcome.Error:
+                                case UnitTestOutcome.Failed:
+                                case UnitTestOutcome.Timeout:
+                                    scope.Span.SetTag(TestTags.Status, TestTags.StatusFail);
+                                    scope.Span.Error = true;
+                                    scope.Span.SetTag(Tags.ErrorMsg, errorMessage);
+                                    scope.Span.SetTag(Tags.ErrorStack, errorStackTrace);
+                                    break;
+                                case UnitTestOutcome.Inconclusive:
+                                case UnitTestOutcome.NotRunnable:
+                                    scope.Span.SetTag(TestTags.Status, TestTags.StatusSkip);
+                                    scope.Span.SetTag(TestTags.SkipReason, errorMessage);
+                                    break;
+                                case UnitTestOutcome.Passed:
+                                    scope.Span.SetTag(TestTags.Status, TestTags.StatusPass);
+                                    break;
                             }
                         }
                     }
@@ -72,7 +110,10 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Testing.MsTestV2
                     if (exception != null)
                     {
                         scope.Span.SetException(exception);
+                        scope.Span.SetTag(TestTags.Status, TestTags.StatusFail);
                     }
+
+                    scope.Dispose();
                 }
             }
